% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.acs.education-package.R
\docType{package}
\name{dc.acs.education-package}
\alias{dc.acs.education}
\alias{dc.acs.education-package}
\title{dc.acs.education: Social Data Commons: Education & Training: ACS Post-HS Eduation}
\description{
Allows user to easily get Post High School Education data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
